+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.1 (August XX, 2016)
 - add a mini slideshow with shortcode to your post or site
+- added upload progress bar
 - added more languages
     - hungarian